+++ test/src/multi_release_test.dart
@@ -22,6 +22,8 @@ void main() {
           tags: null,
           excludeTags: 'bad',
           format: true,
+          ignoreWarnings: false,
+          skipPackages: [],
           useGit: true);
     }, keep: true);
   });